TicTacToe
index.php is the main file for this project which houses the html code
and the links to the CSS and JS.
Scripting.js is the code for the project. This is not a great optimized
code, but it works.
style.css is responsible for all the styling in the project. The
buttons, fonts, positioning, etc.
WinningCombinations is a .txt file I used to keep track what
combinations are winning combinations. The x, y combo are in there,
followed by the ButtonsArray value, followed by the piece of text
explaining what coordinate means what value in the array.